+++ csg-tutorials/propane/imc-ibiintra/run.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+set -e
+
+# IMC iterations
+cmd='csg_inverse --options settings.xml'
+echo "now running: ${cmd}"
+$cmd