+++ xtp/src/libxtp/gwbse/gwbse.cc
@@ -1,7 +1,7 @@
 
 
 /*
- *            Copyright 2009-2022 The VOTCA Development Team
+ *            Copyright 2009-2023 The VOTCA Development Team
  *                       (http://www.votca.org)
  *
  *      Licensed under the Apache License, Version 2.0 (the "License")
@@ -21,7 +21,6 @@
 
 // Third party includes
 #include <boost/algorithm/string.hpp>
-#include <boost/filesystem.hpp>
 #include <boost/format.hpp>
 
 // VOTCA includes
@@ -40,7 +39,6 @@
 #include "votca/xtp/vxc_potential.h"
 
 using boost::format;
-using namespace boost::filesystem;
 using std::flush;
 namespace votca {
 namespace xtp {
@@ -595,7 +593,6 @@ bool GWBSE::Evaluate() {
   XTP_LOG(Log::error, *pLog_)
       << TimeStamp() << " Loaded Auxbasis Set " << auxbasis_name_ << flush;
 
-  // fill auxiliary AO basis by going through all atoms
   orbitals_.SetupAuxBasis(auxbasis_name_);
   AOBasis auxbasis = orbitals_.getAuxBasis();
   XTP_LOG(Log::error, *pLog_) << TimeStamp() << " Filled Auxbasis of size "