+++ xtp/src/libxtp/qmpackages/xtpdft.cc
@@ -1,5 +1,5 @@
 /*
- *            Copyright 2009-2022 The VOTCA Development Team
+ *            Copyright 2009-2023 The VOTCA Development Team
  *                       (http://www.votca.org)
  *
  *      Licensed under the Apache License, Version 2.0 (the "License")
@@ -71,7 +71,13 @@ bool XTPDFT::RunActiveDFT() {
   DFTEngine xtpdft;
   xtpdft.Initialize(options_);
   xtpdft.setLogger(pLog_);
-  bool success = xtpdft.EvaluateActiveRegion(orbitals_);
+
+  if (!externalsites_.empty()) {
+    xtpdft.setExternalcharges(&externalsites_);
+  }
+
+  bool success = xtpdft.EvaluateActiveRegion(orbitals_) &&
+                 xtpdft.EvaluateTruncatedActiveRegion(orbitals_);
   std::string file_name = run_dir_ + "/" + log_file_name_;
   XTP_LOG(Log::error, *pLog_)
       << "Writing embedding result to " << log_file_name_ << flush;