+++ CHANGELOG.rst
@@ -1,8 +1,35 @@
 For more detailed information about the changes see the history of the
-`repository <https://github.com/votca/votca/commits/stable>`__.
+`repository <https://github.com/votca/votca/commits/master>`__.
 
-Version 2023-dev
-================
+Version 2023-rc.2-dev
+=====================
+
+-  fix pyxtp option restart (#1074)
+-  fix Changelog check (#1070)
+-  add python bindings for xtp (#1061)
+-  replaced pointers in MD2QM mapping (#1075)
+-  fixed typo in esp2multipole option (#1076)
+
+Version 2023-rc.2 (released 02.06.23)
+=====================================
+
+-  add note on gmx build for install (#1052)
+-  fix typos in esp2multipole options (#1054)
+-  Used sharedable lock in xtp (#1057)
+-  Make votca tests compile without OpenMP (#1065)
+-  increase tolerance for diabatization test (#1039, #1043)
+-  removed boost::filesystem  (#1041)
+-  clean up C++ defines (#1044)
+-  drop gmx-2022 builds from CI (#1046)
+-  update ccaching in CI (#1048, #1050)
+-  add support for gmx-2023 api (#1059)
+-  updated formatting (#1062)
+-  remove stable branch workflow (#1066)
+-  Refactor changelog actions (#1068)
+-  fix CI build with icpx (#1018)
+
+Version 2023-rc.1 (released 02.02.23)
+=====================================
 
 -  implemented projection-based DFT embedding (#964)
 -  Added LMO energy output and cube file generation (#970)
@@ -10,17 +37,19 @@ Version 2023-dev
 -  unitary optimization for localization of orbitals (#971)
 -  separate option for density max convergence tolerance in qmmm (#987)
 -  Correct ECP charges added in Orca Molden reader (#990)
+-  Implemented truncated basis for embedded DFT (#991)
 -  Implemented ER diabatization for BSE (#988)
 -  update esp++ spce tutorial to py3 (#997)
+-  Truncated GW-in-DFT embedded in static/polar MM implemeted (#1003)
 -  reducing XTP tutorial size (#1020, #1025)
 -  Clean up 3body Stillinger-Weber implement, docs and tutorial (#1012)
 -  fixed options handling in xtp_autogen_mapping (#1024)
 -  Major CMake refactor (#950, #1022)
-
-Version 2022.2 (released XX.12.22)
-==================================
-
--  fix CI build with icpx (#1018)
+-  Fix all doxygen warnings (#1026)
+-  basis and auxbasis are written to QMMM checkpoint file (#1030)
+-  iterations_max can now be zero (#1033)
+-  add an XML check to csg_inverse (#1031)
+-  fix propane/imc and add propane/imc-ibiintra to csg-tutorials (#821)
 
 Version 2022.1 (released 21.12.22)
 ==================================
@@ -37,7 +66,7 @@ Version 2022.1 (released 21.12.22)
 -  fix CI builds with icpx (#984, #1004)
 -  fix CI builds with ESPResSo 4.2.0 (#993)
 -  fix localization issue (#998, #1000)
--  update github actions to new versions (#1007, #1013)
+-  update github actions to new versions (#1007, #1013, #1028)
 
 Version 2022 (released 15.01.22)
 ================================